Moved StreamingLinearAlgorithm to pyspark.mllib.regression
